feat(combobox-web): editorConfig changes, minor refactor, version bump
