+++ packages/pluggableWidgets/combobox-web/CHANGELOG.md
@@ -14,6 +14,10 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 
 -   We made a minor change on the configuration labels.
 
+### Added
+
+-   We added database list feature support.
+
 ## [1.1.3] - 2024-01-22
 
 ### Fixed
