+++ packages/pluggableWidgets/combobox-web/package.json
@@ -1,7 +1,7 @@
 {
   "name": "@mendix/combobox-web",
   "widgetName": "Combobox",
-  "version": "1.1.3",
+  "version": "1.2.0",
   "description": "Configurable Combo box widget with suggestions and autocomplete.",
   "copyright": "© Mendix Technology BV 2024. All rights reserved.",
   "license": "Apache-2.0",
