+++ packages/pluggableWidgets/combobox-web/src/Combobox.editorPreview.tsx
@@ -3,14 +3,11 @@ import { ReferenceValue } from "mendix";
 import { ComponentType, ReactElement, ReactNode, createElement, useMemo } from "react";
 import { ComboboxPreviewProps, OptionsSourceAssociationCustomContentTypeEnum } from "../typings/ComboboxProps";
 import { SingleSelection } from "./components/SingleSelection/SingleSelection";
-import {
-    AssociationSimpleCaptionsProvider,
-    CaptionContent
-} from "./helpers/Association/AssociationSimpleCaptionsProvider";
 import { AssociationOptionsProvider } from "./helpers/Association/AssociationOptionsProvider";
+import { AssociationSimpleCaptionsProvider } from "./helpers/Association/AssociationSimpleCaptionsProvider";
 import { BaseAssociationSelector } from "./helpers/Association/BaseAssociationSelector";
 import { CaptionPlacement, SingleSelector } from "./helpers/types";
-import { getDatasourcePlaceholderText } from "./helpers/utils";
+import { CaptionContent, getDatasourcePlaceholderText } from "./helpers/utils";
 import "./ui/Combobox.scss";
 
 interface PreviewProps {
