+++ packages/pluggableWidgets/combobox-web/src/assets/icons.tsx
@@ -1,7 +1,6 @@
 import classNames from "classnames";
 import { MouseEvent, ReactElement, createElement, Fragment } from "react";
-import { CaptionContent } from "../helpers/Association/AssociationSimpleCaptionsProvider";
-
+import { CaptionContent } from "../helpers/utils";
 export function ClearButton({ size = 14 }): ReactElement {
     return (
         <span className="widget-combobox-icon-container">
