+++ packages/pluggableWidgets/combobox-web/src/helpers/Database/DatabaseCaptionsProvider.tsx
@@ -1,7 +1,8 @@
 import { DynamicValue, ListAttributeValue, ListExpressionValue, ListWidgetValue, ObjectItem } from "mendix";
-import { ReactNode, createElement, MouseEvent, PropsWithChildren, ReactElement } from "react";
+import { ReactNode, createElement } from "react";
 import { OptionsSourceAssociationCustomContentTypeEnum } from "../../../typings/ComboboxProps";
 import { CaptionPlacement, CaptionsProvider } from "../types";
+import { CaptionContent } from "../utils";
 
 interface Props {
     emptyOptionText?: DynamicValue<string>;
@@ -9,27 +10,6 @@ interface Props {
     customContent?: ListWidgetValue | undefined;
     customContentType: OptionsSourceAssociationCustomContentTypeEnum;
     caption?: string;
-}
-
-interface CaptionContentProps extends PropsWithChildren {
-    htmlFor?: string;
-    onClick?: (e: MouseEvent) => void;
-}
-
-export function CaptionContent(props: CaptionContentProps): ReactElement {
-    const { htmlFor, children, onClick } = props;
-    return createElement(htmlFor == null ? "span" : "label", {
-        children,
-        className: "widget-combobox-caption-text",
-        htmlFor,
-        onClick: onClick
-            ? onClick
-            : htmlFor
-            ? (e: MouseEvent) => {
-                  e.preventDefault();
-              }
-            : undefined
-    });
 }
 
 export class DatabaseCaptionsProvider implements CaptionsProvider {
