+++ packages/pluggableWidgets/combobox-web/src/helpers/utils.ts
@@ -1,5 +1,6 @@
 import { ComboboxPreviewProps } from "typings/ComboboxProps";
 import { MultiSelector } from "./types";
+import { PropsWithChildren, ReactElement, createElement } from "react";
 
 export function getSelectedCaptionsPlaceholder(selector: MultiSelector, selectedItems: string[]): string {
     if (selectedItems.length === 0) {
@@ -14,6 +15,27 @@ export function getSelectedCaptionsPlaceholder(selector: MultiSelector, selected
         return "";
     }
     return selectedItems.map(v => selector.caption.get(v)).join(", ");
+}
+
+export interface CaptionContentProps extends PropsWithChildren {
+    htmlFor?: string;
+    onClick?: (e: MouseEvent) => void;
+}
+
+export function CaptionContent(props: CaptionContentProps): ReactElement {
+    const { htmlFor, children, onClick } = props;
+    return createElement(htmlFor == null ? "span" : "label", {
+        children,
+        className: "widget-combobox-caption-text",
+        htmlFor,
+        onClick: onClick
+            ? onClick
+            : htmlFor
+            ? (e: MouseEvent) => {
+                  e.preventDefault();
+              }
+            : undefined
+    });
 }
 
 export function getDatasourcePlaceholderText(args: ComboboxPreviewProps): string {
